Randomize test indices in "load" scheduling
When running tests with --dist=load and n nodes, the scheduler will
statically allocate the tests for each runner: the first 1/n-th of the tests
to the first runner, etc.

The problem with this is that, in practice, slow tests are not evenly
distributed among the test collection. Instead, one should expect modules with
fast tests and other modules with slower tests, which means that most slow
tests will be likely allocated to the same runners.

This commit shuffles the collection of (indices of) tests prior to doing
the static allocation. In this way slow tests will end up more evenly distributed,
sharing the load between more runners.